# master/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Nucleon Forge Axile
  version: 1.0.0

tags:
  - name: Loggers API Controller
    description: The loggers endpoint provides access to the application’s loggers and the configuration of their levels.

  - name: Wallboard API
    description: The endpoints related to wallboard grid rendering

  - name: Beans API
    description: The beans endpoint provides information about the application’s beans.

  - name: Configprops API Controller
    description: The configprops endpoint provides information about the application’s @ConfigurationProperties beans.

  - name: General instance information API
    description: The info endpoint provides general information about the particular Spring Boot instance

  - name: Property Management API
    description: Provides operations for managing Spring properties of application instances.

  - name: Environment API Controller
    description: The env endpoint provides information about the application’s Environment.

  - name: Profile Management API
    description: Provides operations for managing Spring profiles of application instances.

  - name: Caches API Controller
    description: The caches endpoint provides access to the application’s caches.

  - name: API for working with Users
    description: The beans endpoint provides information about the application’s beans.

paths:
######### Users API -> /response/user.yaml -> LoginRequest #########
  /users/login:
    post:
      tags:
        - API for working with Users
      summary: Log-in by the username/password combination
      operationId: login
      parameters:
        - name: loginRequest
          in: query
          description: Request for login
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: './response/user.yaml#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            Authorization:
              description: The JWT token that should be subsequently used for auth purposes
              required: true
              style: simple
              schema:
                type: string
          content:
            '*/*':
              schema:
                type: object
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "401":
          $ref: "./response/error.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./response/error.yaml#/components/responses/Forbidden"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Wallboard API -> /response/wallboard.yaml -> InstancesGridResponse #########
  /applications/grid:
    get:
      tags:
        - Wallboard API
      summary: Returns all instances that are managed by this Axile deployment
      operationId: getInstancesGrid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/wallboard.yaml#/components/schemas/InstancesGridResponse'
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Beans API -> /response/beans.yaml -> BeansFeedResponse #########
  /beans/feed/{instanceId}:
    get:
      tags:
        - Beans API
      summary: Returns beans feed for the given instance
      operationId: getBeansProfile
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/beans.yaml#/components/schemas/BeansFeedResponse'
          links:
            Actuator/Beans:
              description: 'https://docs.spring.io/spring-boot/api/rest/actuator/beans.html'
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Environment API -> /response/environment.yaml -> EnvironmentFeedResponse #########
  /env/feed/{instanceId}:
    get:
      tags:
        - Environment API Controller
      summary: Returns information about the application’s Environment.
      operationId: getEnvironment
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/environment.yaml#/components/schemas/EnvironmentFeedResponse'
          links:
            Spring Boot / Actuator / Environment (env):
              description: 'https://docs.spring.io/spring-boot/api/rest/actuator/env.html'
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Environment API -> /response/environment.yaml -> EnvironmentPropertyResponse #########
  /env/{instanceId}/property/{propertyName}:
    get:
      tags:
        - Environment API Controller
      summary: Returns a specific property of an instance
      operationId: getProperty
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
        - name: propertyName
          in: path
          description: Name of the environment property
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/environment.yaml#/components/schemas/EnvironmentPropertyResponse'
          links:
            Actuator/Environment(env):
              description: 'https://docs.spring.io/spring-boot/api/rest/actuator/env.html'
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Info API -> /response/info.yaml -> InfoResponse #########
  /info/{instanceId}:
    get:
      tags:
        - General instance information API
      summary: Returns general information about the instance.
      operationId: getInfoResponse
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/info.yaml#/components/schemas/InfoResponse'
          links:
            Actuator/Info:
              description: 'https://docs.spring.io/spring-boot/api/rest/actuator/info.html'
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Software API -> /response/software.yaml -> InfoResponse #########
  /software/core/summary:
    get:
      tags:
        - software-api
      operationId: getCoreDistributionGrid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: './response/software.yaml#/components/schemas/DistributionSummaryResponse'

######### Loggers API - GET -> /response/loggers.yaml -> LoggersResponse #########
  '/loggers/{instanceId}':
    get:
      tags:
        - Loggers API Controller
      summary: Returns details of the application’s loggers.
      operationId: getAllLoggers
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/loggers.yaml#/components/schemas/LoggersResponse'
          links:
            Actuator/Loggers:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/loggers.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Loggers API - GET -> /response/loggers.yaml -> LoggerProfileResponse #########
  '/loggers/{instanceId}/logger/{loggerName}':
    get:
      tags:
        - Loggers API Controller
      summary: Returns the details of the requested logger.
      operationId: getLoggerByName
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
        - name: loggerName
          in: path
          description: The name of the logger to find
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/loggers.yaml#/components/schemas/LoggerProfileResponse'
          links:
            Actuator/Loggers:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/loggers.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Loggers API - POST -> No response #########
    post:
      tags:
        - Loggers API Controller
      summary: The request specifies the desired logging level for a logger by its name.
      description: >-
        Suggested logging levels that the user can select to configure the
        logger: OFF, FATAL, ERROR, WARN, INFO, DEBUG, TRACE
      operationId: setLoggingLevelByLoggerName
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
        - name: loggerName
          in: path
          description: The name of the logger to find
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: './response/loggers.yaml#/components/schemas/LogLevelChangeRequest'
        required: true
      responses:
        '200':
          description: OK
          links:
            Actuator/Loggers:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/loggers.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

  ######### Loggers API - POST -> No response #########
  '/loggers/{instanceId}/logger/{loggerName}/clear':
    post:
      tags:
        - Loggers API Controller
      summary: Clears the configured logging level of a logger, reverting it to the global/default setting.
      operationId: clearLoggingLevelByLoggerName
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
        - name: loggerName
          in: path
          description: The name of the logger to find
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          links:
            Actuator/Loggers:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/loggers.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Loggers API - GET -> /response/loggers.yaml -> GroupProfileResponse #########
  '/loggers/{instanceId}/group/{groupName}':
    get:
      tags:
        - Loggers API Controller
      summary: Returns details of the requested logger group.
      operationId: getGroupByName
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
        - name: groupName
          in: path
          description: The name of the logger group to find
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/loggers.yaml#/components/schemas/GroupProfileResponse'
          links:
            Actuator/Loggers:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/loggers.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Loggers API - POST -> No response #########
    post:
      tags:
        - Loggers API Controller
      summary: The request specifies the desired logging level for a logger group by its name.
      description: >-
        Suggested logging levels that the user can select to configure the
        logger: OFF, FATAL, ERROR, WARN, INFO, DEBUG, TRACE
      operationId: setLoggingLevelByGroupName
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
        - name: groupName
          in: path
          description: The name of the logger group to find
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: './response/loggers.yaml#/components/schemas/LogLevelChangeRequest'
        required: true
      responses:
        '200':
          description: OK
          links:
            Actuator/Loggers:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/loggers.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Profile Management API - POST -> /response/profile-management.yaml -> ProfileUpdateResponse #########
  '/profile-management/{instanceId}':
    post:
      tags:
        - Profile Management API
      summary: Replaces active profiles for the given application instance.
      operationId: replaceProfile
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
      requestBody:
        content:
          application/json:
            schema:
              $ref: './response/profile-management.yaml#/components/schemas/ProfileUpdatedRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/profile-management.yaml#/components/schemas/ProfileUpdateResponse'
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Configprops API  -> /response/configprops.yaml -> ConfigpropsFeedResponse #########
  '/configprops/feed/{instanceId}':
    get:
      tags:
        - Configprops API Controller
      summary: Returns all @ConfigurationProperties beans of the application.
      operationId: getBeansFeedProfile
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/configprops.yaml#/components/schemas/ConfigpropsFeedResponse'
          links:
            Actuator/Configuration Properties:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/configprops.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

######### Configprops API  -> /response/configprops.yaml -> ConfigpropsByPrefixResponse #########
  '/configprops/{instanceId}/beans/{prefix}':
    get:
      tags:
        - Configprops API Controller
      summary: Returns the @ConfigurationProperties beans of a given instance with given prefix
      operationId: getBeanByPrefixProfile
      parameters:
        - $ref: "#/components/parameters/InstanceIdParam"
        - name: prefix
          in: path
          description: The prefix of @ConfigurationProperties beans to find
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './response/configprops.yaml#/components/schemas/ConfigpropsByPrefixResponse'
          links:
            Actuator/Configuration Properties:
              description: https://docs.spring.io/spring-boot/api/rest/actuator/configprops.html
        "400":
          $ref: "./response/error.yaml#/components/responses/BadRequest"
        "500":
          $ref: "./response/error.yaml#/components/responses/InternalError"

components:
  parameters:
    InstanceIdParam:
      name: instanceId
      in: path
      description: Application Instance ID
      required: true
      schema:
        type: string